Moves Login component to views with minor refactor
