+++ client/src/views/login/style.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components'
+
+export const CenterBlock = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100vh;
+`
+
+export const CenterBox = styled.div`
+  width: 320px;
+`
+
+export const StyledForm = styled.form`
+  margin-bottom: 20px;
+`
